requirements.txt
asgiref==3.11.0
Django==6.0
django-crispy-forms==2.5
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
PyMySQL==1.1.2
python-decouple==3.8
sqlparse==0.5.5
whitenoise==6.11.0
